backend/app/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pytesseract
pillow
pdf2image
reportlab
numpy
openai
python-dotenv
